Update README.md
Added my favorite movie/series such as:

The Land Before Time - An orphaned brontosaurus teams up with other young dinosaurs in order to reunite with their families in a valley.
Jurassic Park - A pragmatic paleontologist touring an almost complete theme park on an island in Central America is tasked with protecting a couple of kids after a power failure causes the park's cloned dinosaurs to run loose.
Huge advancements in scientific technology have enabled a mogul to create an island full of living dinosaurs. John Hammond has invited four individuals, along with his two grandchildren, to join him at Jurassic Park. But will everything go according to plan? A park employee attempts to steal dinosaur embryos, critical security systems are shut down and it now becomes a race for survival with dinosaurs roaming freely over the island.
Scorpion - An eccentric genius forms an international network of super-geniuses to act as the last line of defense against the complicated threats of the modern world.